new encoder configuration files
