+++ gripper/firmware/motor_hardware.h
@@ -12,7 +12,6 @@ extern "C" {
 
 extern SPI_HandleTypeDef hspi2;
 extern TIM_HandleTypeDef htim1;
-extern TIM_HandleTypeDef htim2;
 extern TIM_HandleTypeDef htim3;
 extern TIM_HandleTypeDef htim7;
 extern DAC_HandleTypeDef hdac1;
@@ -24,7 +23,6 @@ HAL_StatusTypeDef initialize_spi();
 void initialize_timer(motor_interrupt_callback callback);
 
 void initialize_dac();
-void initialize_enc();
 
 void update_pwm(uint32_t freq, uint32_t duty_cycle);
